gh-actions: update flake.lock on push too
Updating it only when the action is triggered by a repository dispatch
(channel update) would make sense if the build job skipped recreating
the lock file on push, but it recreates it every time. So the current
behavior is that packages are always built against the latest nixpkgs on
push, and therefore the flake.lock should be updated (when applicable)
as well on push.